move get_language to public as it is used by plugins
