+++ src/goose/_internal/toolkit/developer.py
@@ -11,7 +11,7 @@ from rich.table import Table
 from rich.text import Text
 
 from goose.toolkit import Toolkit, tool
-from .utils import get_language
+from goose.utils.file_language import get_language
 
 
 def keep_unsafe_command_prompt(command: str) -> PromptType:
